use upsert to update existent users or insert new ones

+++ api/functions/updateDbUsers.js
@@ -1,5 +1,9 @@
 export const updateDbUsers = async (usersFile, db) => {
 	try {
+		if (!usersFile) {
+			return { success: false };
+		}
+
 		await db.query("BEGIN");
 
 		await db.query("DELETE FROM all_users");
@@ -7,25 +11,31 @@ export const updateDbUsers = async (usersFile, db) => {
 		const activeUsers = usersFile.filter((user) => !user.deleted);
 
 		if (activeUsers.length === 0) {
-			db.query("ROLLBACK");
+			await db.query("ROLLBACK");
 			return {
 				success: false,
 				message: "active user is empty.something is wrong",
 			};
 		}
 
-		// batch insert query
+		// Batch upsert query
 		const insertQuery = `
             INSERT INTO all_users (user_id, display_name, display_name_normalised, is_admin, email)
             VALUES ${activeUsers.map((_, i) => `($${i * 5 + 1}, $${i * 5 + 2}, $${i * 5 + 3}, $${i * 5 + 4}, $${i * 5 + 5})`).join(", ")}
+            ON CONFLICT (user_id) 
+            DO UPDATE SET 
+                display_name = EXCLUDED.display_name,
+                display_name_normalised = EXCLUDED.display_name_normalised,
+                is_admin = EXCLUDED.is_admin,
+                email = EXCLUDED.email
         `;
 
 		const queryValues = activeUsers.flatMap((user) => [
-			user.id,
-			user.profile.display_name,
-			user.profile.display_name_normalized,
-			user.is_admin,
-			user.profile.email,
+			user.id ?? "unkown",
+			user.profile.display_name ?? "",
+			user.profile.display_name_normalized ?? "",
+			user.is_admin ?? false,
+			user.profile.email ?? "",
 		]);
 
 		// Insert all users in a single query
